CARL ResourceManager: Move `resource::types::resources` underneath `resource::manager`.
