+++ opendut-carl/src/manager/cluster_manager/list_cluster_peers.rs
@@ -1,4 +1,4 @@
-use crate::resource::types::resources::Resources;
+use crate::resource::manager::Resources;
 use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 use opendut_model::cluster::{ClusterDescriptor, ClusterId};
