+++ opendut-carl/src/manager/peer_manager/list_peer_member_states.rs
@@ -1,5 +1,5 @@
 use crate::manager::cluster_manager;
-use crate::resource::types::resources::Resources;
+use crate::resource::manager::Resources;
 use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 use opendut_model::cluster::ClusterId;
